requirements.txt
asyncio-throttle==1.0.2
autopep8==1.5.7
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
idna==2.10
numpy==1.20.3
pandas==1.2.4
pycodestyle==2.7.0
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
toml==0.10.2
urllib3==1.26.5
